requirements.txt
imutils==0.5.4
numpy==1.21.1
opencv-python==4.5.3.56
pdf2image==1.16.0
Pillow==8.3.1
pytesseract==0.3.8
